Declared GNU Affero GPL License v3.0
This is the strongest copyleft license made by the Free Software Modification. It has all the terms and conditions of the original GNU GPL v3.0 license, with one additional requirement: It has one added requirement: if you run a modified program on a server and let other users communicate with it there, your server must also allow them to download the source code corresponding to the modified version running there.